refactor: api/authorization/ を pkg/authz/ と graph/ に再配置
- Authorizer interface・StaticAuthorizer・RoleCache を api/pkg/authz/ に移動
  - 認可（authz）専用パッケージとして auth と対になる命名
  - DB・フレームワーク非依存の純粋ロジックは pkg/ に置く方針に合わせた
- @hasPermission ディレクティブハンドラを api/graph/directives.go に移動
  - gqlgen の慣習通り graph パッケージ直下に配置
- api/authorization/ ディレクトリを削除
