+++ api/pkg/authz/role_cache.go
@@ -1,4 +1,4 @@
-package authorization
+package authz
 
 import (
 	"sync"
